ISSUE-6: Remove unnecessary database/sql import
It was added when a column of null-type was detected regardless if it
is a time value or not. If the table has no null values beside the time
null value then the import is not used and can be removed safely.

Added new function to wrap the generation of import statements and used a
new struct to hold some information about a column when it gets processed.