ISSUE-22: Introduce Makefile
SQLite3 will be disabled nby default. It can be built/enabled via
built-tag `sqlite3`. The Makefile exists for convenience. If the new
type is used tables-to-go will fail with the message driver not found.
Therefore the separate sqlite_driver.go file is introduced.

Moved the default username from Settings to Postgres implementation.

Created unit tests for DSN methods of Postgres and SQLite.